@@ -1,6 +1,6 @@
 import { Component, ViewEncapsulation } from '@angular/core';
 import { GraphModule } from 'src/app/classes/graphModule';
-import { MatBottomSheet } from '@angular/material';
+import { MatBottomSheet } from '@angular/material/bottom-sheet';
 import { CharacterService } from 'src/app/character.service';
 
 @Component({
